Works in multiplayer
