+++ assets/prefabs/blade.prefab
@@ -7,7 +7,7 @@
     "kinematic": true,
     "mass" : 100,
     "collisionGroup" : "engine:sensor",
-    "collidesWith" : ["engine:sensor"]
+    "collidesWith" : ["engine:sensor", "engine:character"]
   },
   "DamagePlayer" : {},
   "Network" :{}
